drivers:axonfabric: Add support for ctrlreg
Add support for accessing the control register in the fabric. The purpose at this point
is to enable support for the shared_sysconfig_disable bit to be turned off in userspace
using the sysfs. This bit, when set to 1, disables the I/O pins on the fabric which are
shared with the SPI slave sysconfig interface on the fabric. This avoids the problem
where the fabric could be driving the SPI_B_MISO pin from both the fabric IO pin and
the SPI sysconfig MISO pin.