sc16is7xx: Fix for "Unexpected interrupt: 8" (backported from mainline kernel 5.1)
Original patch is exist in tag v5.0-rc1 from linux kernel mainline

commit 30ec514d440cf2c472c8e4b0079af2c731f71a3e
Author: Phil Elwell <phil@raspberrypi.org>
Date:   Wed Sep 12 15:31:56 2018 +0100
Title: sc16is7xx: Fix for "Unexpected interrupt: 8"

The SC16IS752 has an Enhanced Feature Register which is aliased at the
same address as the Interrupt Identification Register; accessing it
requires that a magic value is written to the Line Configuration
Register. If an interrupt is raised while the EFR is mapped in then
the ISR won't be able to access the IIR, leading to the "Unexpected
interrupt" error messages.

Avoid the problem by claiming a mutex around accesses to the EFR
register, also claiming the mutex in the interrupt handler work
item (this is equivalent to disabling interrupts to interlock against
a non-threaded interrupt handler).

See: https://github.com/raspberrypi/linux/issues/2529

Signed-off-by: Phil Elwell <phil@raspberrypi.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>