wm8960: Keep freq_in consistent in device data
When using a fixed mclk, which uses the onboard PLL to generate the sysclk,
set_dai_pll is called when the device playback or recording is done. If we
allow this function to set freq_in in the device data, the next time the codec
is engaged the freq_in will be set to zero (this happens when the pll is
being turned off). Do not allow this function to set freq_in in the
device data.