ARM: dts: axon-imx6: Add mSD pwr ctrl, update FEC
Add SD power control to USDHC1, as designed into axon-pi rev 0.3.
Modify FEC to same configuration as pico-imx6, to fix issue with
PHY configuration. The symptom of the problem with PHY was evident
in the increased number of errors in received frames, as reported
by ifconfig. In addtion, DHCP fails to claim an address or other
network configuration.