ARM64: tn_imx8_defconfig: Fix SDMA firmware load issue
Make SDMA driver a loadable module instead of built-in. This is a workaround
for an issue with SDMA firmware loading. Without this workaround, these errors
are seen with PICO-IMX8MQ:
[    1.157317] imx-sdma 30bd0000.sdma: no iram assigned, using external mem
[    1.163605] imx-sdma 30bd0000.sdma: Falling back to user helper
[    1.168088] imx-sdma 302c0000.sdma: no iram assigned, using external mem
[    5.407718] imx-sdma 302c0000.sdma: external firmware not found, using ROM firmware
[    5.413315] imx-sdma 30bd0000.sdma: external firmware not found, using ROM firmware

With the workaround, the firmware is loaded from the rootfs:
[    3.822638] imx-sdma 30bd0000.sdma: no iram assigned, using external mem
[    3.856903] imx-sdma 302c0000.sdma: no iram assigned, using external mem
[    3.910708] imx-sdma 30bd0000.sdma: loaded firmware 4.4
[    3.919143] imx-sdma 302c0000.sdma: loaded firmware 4.4

If SDMA firmware is not loaded, then errors can occur when attempting to
use UARTs, SPI, and other interfaces which make use of SDMA.