drivers: mfd: axonfabric: Set gpio can_sleep flag to false
The ili9881c driver only calls gpiod_set_value. When this happens, if the
can_sleep flag is TRUE in the gpio driver, the gpiod driver will issue
a warning that the driver should be calling the equivalent _can_sleep()
function. The warnings appear as exceptions in the system log. This
makes the can_sleep flag false in the axonfabric gpio driver, so the
messages will not appear. The ili9881c driver should be fixed to call
the appropriate function, so this is a TODO for a later time.