ARM64:dts:axon-imx8mm: Fix BT UART
In version A1 on axon-imx8mm, the Bluetooth UART signals connect through the
Axon fabric (FPGA) in the same way as on axon-imx6. This patch changes the
pinmux settings for the BT UART (actually fixes them).

Also, it changes the axonfabric settings to enable the BT UART passthrough,
and extends the post-reset period to 200 ms for the BT_REG_ON reset. This is
because in some cases, it can take longer for the BT radio to initialize
after toggling BT_REG_ON.