ARM64:dts:imx8mm-axon: Update pins for rev A1, fix baseboard sd, move SD to SOM DT
On rev A1, in order to make all SAI1 pins available for use to the axon fabric
we moved some pins around. This required to change the definitions for the
bitbang i2c bus (brd_conf).

In addition, the VMMC regulator (SD_RESET_B) enable changed, so corrected this
and moved all the baseboard SD card definitions to the SOM DT.

Also, fixed the baseboard SD card detect signal. It should be a GPIO.