ARM64:dts:imx8mm-axon-pi: remove LINPUT3 and mipi touch reset
Remove the LINPUT3 from the WM9860 configuration as we are using a
single-ended mic by default.

Remove the MIPI touch reset signal as we do not use that by default
on AXON-PI. We reset the touch panel with the main board reset signal.

Make F_GPIO_X1_P70 an axon fabric I2C GPIO (not passthrough) for the
user LED on the AXON PI baseboard.