ARM64:dts:imx8mm-axon: add loopbacktest device tree
It is useful to test the connections between the SOC and the FPGA.
For this purpose we have a special device tree which both disables
a number of interfaces that claim I/O pins that need to be tested,
and adds a pinctrl group to enable the axonfabric device to claim
all of the relevant pins.

It should be noted that this only tests the I/Os that can be
programmed via the SOC over the I2C interface control connection to
a loaded FPGA running axonfabric firmware. Not all of the pins are
testable (e.g. the spare GPIOs, interrupt inputs) in this fashion.