arm32:dts:imx6qdl-axon: Changes for gpio driver, spi, otg
Migrate the axonfabric gpio driver to separate it from the pinmux driver.
This allows it to be a consumer of the pinmux configuration.

Make the SPI interface MISO signals HIGH-Z outputs to avoid conflicts with
axonfabric MISO signals. It was noticed that, despite being in master
mode, the MISO signals were being driven by the SOC. This causes contention
on the signal lines as the fabric (which passes through the slave MISO)
wants to drive the signal as well.

Clear up some OTG control signal definitions.