drivers:mfd:axonfabric: Add reset code
During warm-boot testing, it was discovered that the module would intermittenly
hang upon boot, and be unable to communicate with the axonfabric firmware in the
AXON fpga. The driver already had a reset GPIO assigned and allocated, but no
reset was being issued to the FPGA upon driver init. Add code to reset the
FPGA (reset to internal FPGA firmware registers) during driver init. Reset GPIO
is freed upon driver removal.