arm64: dts: imx8mq-pico: add VL101-12880YL-C13(M101GWWF R0) panel support
It's for pico-wizard baseboard, due to different panel as previous 10"
panel, so need retune mipi-to-lvds chip timing and relate registers,
note that this timing need compatible with previous 10" panel.

Otherwise, add a power-supply regulator from dsi power pin is necessary,
so add to mipi-to-lvds node. Also add an i2c base exc3000 touchscreen
for this panel.